# dist/api.yml
...
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/BGU-Exercise/Assignment3.1/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://virtserver.swaggerhub.com/tzah4530/exercise/1.0.0
  description: SwaggerHub API Auto Mocking
...
paths:
...
  /players/{playerId}:
    get:
      tags:
      - players
      summary: player preview
      description: player information
      parameters:
      - name: playerId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: id of the player or coach
      responses:
        "200":
          description: player information sent sucessfuly.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/player_choach_personalPage'
        "404":
          description: player doesn't exist.
  /players/search/{playerName}:
    get:
      tags:
      - players
      summary: players preview
      description: players information by Name.
      parameters:
      - name: playerName
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: Name or part of the player name.
      responses:
        "200":
          description: player / coach information sent sucessfuly.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/player_choach_personalPage'
        "404":
          description: player / coach doesn't exist.
  /team/teamFullDetails:
    get:
      tags:
      - team
      summary: All teams
      description: List of all the teams
      responses:
        "200":
          description: All teams information sent sucessfuly.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/team-allteams'
        "404":
          description: No teams exist in system.
  /team/teamFullDetails/{teamId}:
    get:
      tags:
      - team
      summary: team preview
      description: team information
      parameters:
      - name: teamId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: id of the team
      responses:
        "200":
          description: team information sent sucessfuly.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/team-teamPreview'
        "404":
          description: team doesn't exist.
  /team/teamFullDetails/search/{teamName}:
    get:
      tags:
      - team
      summary: team preview
      description: team information
      parameters:
      - name: teamName
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: name of the team.
      responses:
        "200":
          description: team information sent sucessfuly.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/team-teamPreview'
        "404":
          description: team doesn't exist.
components:
  schemas:
...
    team-allteams:
      type: array
      items:
        $ref: '#/components/schemas/teamallteams_inner'
    team-teamPreview:
      type: object
      properties:
        team_id:
          type: integer
          description: team id
          example: 21
        team_name:
          type: string
          description: name of team
          example: Manchester United
        team_tag:
          type: string
          description: short name of the team
          example: REA
        team_players:
          type: array
          description: team players.
          items:
            $ref: '#/components/schemas/teamteamPreview_team_players'
        coach:
          type: integer
          description: team coach id
          example: 25
        coach_name:
          type: string
          description: team coach name
          example: Zinedine Zidane
    player_choach_personalPage:
      type: object
      properties:
        id:
          type: integer
          description: player id
          example: 5627
        name:
          type: string
          description: player name.
          example: Mikkel Andersen
        common_name:
          type: string
          description: player common name.
          example: M. Andersen
        position:
          type: string
          description: position of player in team.
          example: "1"
        team_name:
          type: string
          description: team belong name.
          example: Midtjylland
        nationality:
          type: string
          description: national of player.
          example: Denmark
        birthdate:
          type: string
          description: birthday of player.
          example: 17/12/1988
        birthcountry:
          type: string
          description: birth country.
          example: Denmark
        height:
          type: string
          description: player height.
          example: 196 cm
        weight:
          type: string
          description: player weight.
          example: 91 kg
        image:
          type: string
          description: link of the player photo.
          example: https://cdn.sportmonks.com/images/soccer/players/27/5627.png
...
    teamallteams_inner:
      properties:
        team_id:
          type: integer
          description: team id
          example: 21
        team_name:
          type: string
          description: name of team
          example: Manchester United
    teamteamPreview_team_players:
      properties:
        player_id:
          type: integer
          description: team player id
          example: 4
        player_name:
          type: string
          description: team player name
          example: Christiano Ronaldo
...